+++ components/opc-ua/src/Microsoft.Azure.IIoT.OpcUa.Publisher/tests/Microsoft.Azure.IIoT.OpcUa.Publisher.Tests.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-   </PropertyGroup>
+  </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Autofac.Extras.Moq" Version="5.0.1" />
     <PackageReference Include="AutoFixture" Version="4.11.0" />
