+++ components/opc-ua/src/Microsoft.Azure.IIoT.OpcUa.Registry/tests/Deploy/IoTHubDiscovererDeploymentTests.cs
@@ -3,13 +3,13 @@
 //  Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
 // ------------------------------------------------------------
 
-namespace Microsoft.Azure.IIoT.OpcUa.Publisher.Tests.Deploy {
+namespace Microsoft.Azure.IIoT.OpcUa.Registry.Tests.Deploy {
     using Autofac;
     using Autofac.Extras.Moq;
     using Microsoft.Azure.IIoT.Deploy.Runtime;
     using Microsoft.Azure.IIoT.Hub;
     using Microsoft.Azure.IIoT.Hub.Models;
-    using Microsoft.Azure.IIoT.OpcUa.Publisher.Deploy;
+    using Microsoft.Azure.IIoT.OpcUa.Registry.Deploy;
     using Microsoft.Azure.IIoT.Serializers;
     using Microsoft.Azure.IIoT.Serializers.NewtonSoft;
     using Microsoft.Extensions.Configuration;
@@ -20,12 +20,12 @@ namespace Microsoft.Azure.IIoT.OpcUa.Publisher.Tests.Deploy {
     using Xunit;
 
     /// <summary>
-    /// Test to check layered deployments that are generated by IoTHubPublisherDeployment
+    /// Test to check layered deployments that are generated by IoTHubDiscovererDeployment
     /// </summary>
-    public class IoTHubPublisherDeploymentTest {
+    public class IoTHubDiscovererDeploymentTests {
 
         [Fact]
-        public async Task PublisherDeploymentRoutesTestAsync() {
+        public async Task DiscoveryDeploymentRoutesTestAsync() {
 
             IList<ConfigurationModel> configurationModelList = new List<ConfigurationModel>();
 
@@ -37,30 +37,26 @@ namespace Microsoft.Azure.IIoT.OpcUa.Publisher.Tests.Deploy {
                 .Returns((ConfigurationModel configuration, bool forceUpdate, CancellationToken ct) => Task.FromResult(configuration));
 
             using (var mock = Setup(ioTHubConfigurationServicesMock)) {
-                var publisherDeploymentService = mock.Create<IoTHubPublisherDeployment>();
-                await publisherDeploymentService.StartAsync();
+                var discoveryDeploymentService = mock.Create<IoTHubDiscovererDeployment>();
+                await discoveryDeploymentService.StartAsync();
 
                 Assert.Equal(2, configurationModelList.Count);
 
                 {
                     // Check routes of layered deployment for Linux
                     var configurationModel = configurationModelList[0];
-                    Assert.Equal("__default-opcpublisher", configurationModel.Id);
+                    Assert.Equal("__default-discoverer-linux", configurationModel.Id);
                     Assert.Equal(2, configurationModel.Content.ModulesContent.Count);
-                    Assert.Equal("FROM /messages/modules/publisher/* INTO $upstream",
-                        configurationModel.Content.ModulesContent["$edgeHub"]["properties.desired.routes.publisherToUpstream"]);
-                    Assert.Equal("FROM /messages/* WHERE NOT IS_DEFINED($connectionModuleId) INTO $upstream",
-                        configurationModel.Content.ModulesContent["$edgeHub"]["properties.desired.routes.leafToUpstream"]);
+                    Assert.Equal("FROM /messages/modules/discovery/* INTO $upstream",
+                        configurationModel.Content.ModulesContent["$edgeHub"]["properties.desired.routes.discoveryToUpstream"]);
                 }
                 {
                     // Check routes of layered deployment for Windows
                     var configurationModel = configurationModelList[1];
-                    Assert.Equal("__default-opcpublisher-windows", configurationModel.Id);
+                    Assert.Equal("__default-discoverer-windows", configurationModel.Id);
                     Assert.Equal(2, configurationModel.Content.ModulesContent.Count);
-                    Assert.Equal("FROM /messages/modules/publisher/* INTO $upstream",
-                        configurationModel.Content.ModulesContent["$edgeHub"]["properties.desired.routes.publisherToUpstream"]);
-                    Assert.Equal("FROM /messages/* WHERE NOT IS_DEFINED($connectionModuleId) INTO $upstream",
-                        configurationModel.Content.ModulesContent["$edgeHub"]["properties.desired.routes.leafToUpstream"]);
+                    Assert.Equal("FROM /messages/modules/discovery/* INTO $upstream",
+                        configurationModel.Content.ModulesContent["$edgeHub"]["properties.desired.routes.discoveryToUpstream"]);
                 }
             }
         }
@@ -87,32 +83,17 @@ namespace Microsoft.Azure.IIoT.OpcUa.Publisher.Tests.Deploy {
             configuration["Docker:ImagesTag"] = "4.5.6.7";
 
             using (var mock = Setup(ioTHubConfigurationServicesMock, configuration)) {
-                var publisherDeploymentService = mock.Create<IoTHubPublisherDeployment>();
+                var discoveryDeploymentService = mock.Create<IoTHubDiscovererDeployment>();
                 var jsonserializer = mock.Container.Resolve<IJsonSerializer>();
 
-                await publisherDeploymentService.StartAsync();
+                await discoveryDeploymentService.StartAsync();
 
                 Assert.Equal(2, configurationModelList.Count);
 
                 {
                     // Check details of docker image in layered deployment for Linux
                     var configurationModel = configurationModelList[0];
-                    Assert.Equal("__default-opcpublisher", configurationModel.Id);
-                    Assert.Equal(2, configurationModel.Content.ModulesContent.Count);
-
-                    var registryCredentials = (Newtonsoft.Json.Linq.JObject) configurationModel.Content.ModulesContent["$edgeAgent"]["properties.desired.runtime.settings.registryCredentials.custom"];
-                    Assert.Equal("custom.azurecr.io", registryCredentials.Value<string>("address"));
-                    Assert.Equal("dPassword", registryCredentials.Value<string>("password"));
-                    Assert.Equal("dUser", registryCredentials.Value<string>("username"));
-
-                    var module = (Newtonsoft.Json.Linq.JObject)configurationModel.Content.ModulesContent["$edgeAgent"]["properties.desired.modules.publisher"];
-                    Assert.Equal("custom.azurecr.io/customNamespace/iotedge/opc-publisher:4.5.6.7",
-                        module["settings"].Value<string>("image"));
-                }
-                {
-                    // Check details of docker image in layered deployment for Windows
-                    var configurationModel = configurationModelList[1];
-                    Assert.Equal("__default-opcpublisher-windows", configurationModel.Id);
+                    Assert.Equal("__default-discoverer-linux", configurationModel.Id);
                     Assert.Equal(2, configurationModel.Content.ModulesContent.Count);
 
                     var registryCredentials = (Newtonsoft.Json.Linq.JObject)configurationModel.Content.ModulesContent["$edgeAgent"]["properties.desired.runtime.settings.registryCredentials.custom"];
@@ -120,8 +101,23 @@ namespace Microsoft.Azure.IIoT.OpcUa.Publisher.Tests.Deploy {
                     Assert.Equal("dPassword", registryCredentials.Value<string>("password"));
                     Assert.Equal("dUser", registryCredentials.Value<string>("username"));
 
-                    var module = (Newtonsoft.Json.Linq.JObject)configurationModel.Content.ModulesContent["$edgeAgent"]["properties.desired.modules.publisher"];
-                    Assert.Equal("custom.azurecr.io/customNamespace/iotedge/opc-publisher:4.5.6.7",
+                    var module = (Newtonsoft.Json.Linq.JObject)configurationModel.Content.ModulesContent["$edgeAgent"]["properties.desired.modules.discovery"];
+                    Assert.Equal("custom.azurecr.io/customNamespace/iotedge/discovery:4.5.6.7",
+                        module["settings"].Value<string>("image"));
+                }
+                {
+                    // Check details of docker image in layered deployment for Windows
+                    var configurationModel = configurationModelList[1];
+                    Assert.Equal("__default-discoverer-windows", configurationModel.Id);
+                    Assert.Equal(2, configurationModel.Content.ModulesContent.Count);
+
+                    var registryCredentials = (Newtonsoft.Json.Linq.JObject)configurationModel.Content.ModulesContent["$edgeAgent"]["properties.desired.runtime.settings.registryCredentials.custom"];
+                    Assert.Equal("custom.azurecr.io", registryCredentials.Value<string>("address"));
+                    Assert.Equal("dPassword", registryCredentials.Value<string>("password"));
+                    Assert.Equal("dUser", registryCredentials.Value<string>("username"));
+
+                    var module = (Newtonsoft.Json.Linq.JObject)configurationModel.Content.ModulesContent["$edgeAgent"]["properties.desired.modules.discovery"];
+                    Assert.Equal("custom.azurecr.io/customNamespace/iotedge/discovery:4.5.6.7",
                         module["settings"].Value<string>("image"));
                 }
             }
@@ -165,12 +161,13 @@ namespace Microsoft.Azure.IIoT.OpcUa.Publisher.Tests.Deploy {
                         .Setup(e => e.CreateOrUpdateConfigurationAsync(It.IsAny<ConfigurationModel>(), true, CancellationToken.None))
                         .Returns((ConfigurationModel configuration, bool forceUpdate, CancellationToken ct) => Task.FromResult(configuration));
                     builder.RegisterMock(ioTHubConfigurationServices);
-                } else {
+                }
+                else {
                     builder.RegisterMock(ioTHubConfigurationServicesMock);
                 }
 
-                builder.RegisterType<IoTHubPublisherDeployment>()
-                    .AsSelf();;
+                builder.RegisterType<IoTHubDiscovererDeployment>()
+                    .AsSelf(); ;
             });
             return mock;
         }
