+++ components/opc-ua/src/Microsoft.Azure.IIoT.OpcUa.Registry/tests/Deploy/IoTHubMetricsCollectorDeploymentTests.cs
@@ -22,7 +22,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Registry.Tests.Deploy {
     /// <summary>
     /// Test to check layered deployments that are generated by IoTHubMetricsCollectorDeployment
     /// </summary>
-    public class IoTHubMetricsCollectorDeploymentTest {
+    public class IoTHubMetricsCollectorDeploymentTests {
 
         [Fact]
         public async Task MetricsCollectorDeploymentRoutesTestAsync() {
